+++ src/node/mod.rs
@@ -21,8 +21,8 @@ mod channel_splitter;
 pub use channel_splitter::*;
 mod channel_merger;
 pub use channel_merger::*;
-mod constant_src;
-pub use constant_src::*;
+mod constant;
+pub use constant::*;
 mod panner;
 pub use panner::*;
 mod analyzer;
